printing scan status after scan summary
