+++ cli/src/semgrep/commands/scan.py
@@ -903,6 +903,7 @@ def scan(
                     output_extra,
                     shown_severities,
                     _dependencies,
+                    target_manager
                 ) = semgrep.semgrep_main.main(
                     core_opts_str=core_opts,
                     dump_command_for_core=dump_command_for_core,
@@ -950,6 +951,7 @@ def scan(
                 severities=shown_severities,
                 print_summary=True,
                 engine_type=engine_type,
+                target_manager=target_manager,
             )
 
             run_has_findings = any(filtered_matches_by_rule.values())
